README.md: update the text about the GitBook Markdown doc format!
Note that the link here don't make note of the fact that the GitBook format
spec is hosted in this very repository. We want people to use this repo's
README.md as a template for other documents, so we should avoid including
anything that might become confusing after copy/pasting to another repo.